refactor(gamebook): move 4 EF configs to canonical path (A0.1)
Pre-existing EF configurations for the 4 gamebook entities lived under
BoundedContexts/SessionTracking/Infrastructure/Persistence/Configurations/ —
move them to the canonical Infrastructure/EntityConfigurations/SessionTracking/
path expected by the gamebook multi-book generalization plan (Task A0.1).

Logical content preserved verbatim:
- GamebookCampaignSession: GamebookProgress VO → jsonb via System.Text.Json
- GamebookPhotoArtifact: IReadOnlyList<GamebookSegment> → jsonb via System.Text.Json
- TranslatedParagraph: AppliedGlossaryTerms string[] → text[] native
- GamebookGlossaryEntry: simple property mapping

Column names, indices, query filters all preserved. Migration drift check
generated an empty migration (model snapshot identical) — no schema change.

No code consumers reference the moved config class names by symbol.

Refs: docs/superpowers/plans/2026-05-19-gamebook-multi-book-generalization.md
