+++ apps/api/src/Api/Infrastructure/EntityConfigurations/SessionTracking/GamebookGlossaryEntryEntityConfiguration.cs
@@ -2,14 +2,16 @@ using Api.BoundedContexts.SessionTracking.Domain.Entities;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
 
-namespace Api.BoundedContexts.SessionTracking.Infrastructure.Persistence.Configurations;
+namespace Api.Infrastructure.EntityConfigurations.SessionTracking;
 
 /// <summary>
 /// EF Core configuration for GamebookGlossaryEntry.
 /// Uses direct domain entity mapping (no separate persistence entity).
+/// Phase A0.1 (2026-05-19): moved from BoundedContexts/SessionTracking/Infrastructure/Persistence/Configurations
+/// to canonical Infrastructure/EntityConfigurations location to align with the rest of the codebase.
 /// Iter 1.B — Libro Game Nanolith dogfood demo.
 /// </summary>
-internal sealed class GamebookGlossaryEntryConfiguration : IEntityTypeConfiguration<GamebookGlossaryEntry>
+internal class GamebookGlossaryEntryEntityConfiguration : IEntityTypeConfiguration<GamebookGlossaryEntry>
 {
     public void Configure(EntityTypeBuilder<GamebookGlossaryEntry> builder)
     {
